Update Jenkinsfile of Feature-A 01

+++ Jenkinsfile
@@ -45,7 +45,16 @@ pipeline {
                 sh '''
                     docker rm -f healthcheck || true
                     docker run -d --name healthcheck hello-vince:latest
-                    sleep 5
+
+                    # wait up to 30 seconds for nginx
+                    for i in {1..15}; do
+                    if docker exec healthcheck wget -qO- http://localhost >/dev/null 2>&1; then
+                        echo "Container healthy"
+                        break
+                    fi
+                    sleep 2
+                    done
+
                     docker exec healthcheck wget -qO- http://localhost
                     docker rm -f healthcheck
                 '''
